+++ examples/stitching.py
@@ -32,9 +32,10 @@ results, (rca_cl, lca_cl, ao_cl) = mm.label_geometry(
     path_centerline_lca="../data/centerline_lca.csv",
     bounding_sphere_radius_mm_rca=3.0,
     bounding_sphere_radius_mm_lca=3.0,
-    n_points_intramural=100,
-    anomalous_rca=True,
-    anomalous_lca=False,
+    n_points_takeoff_rca=100,
+    n_points_takeoff_lca=100,
+    acute_takeoff_rca=True,
+    acute_takeoff_lca=False,
     control_plot=False,
 )
 
@@ -167,9 +168,10 @@ results, (rca_cl, lca_cl, ao_cl) = mm.label_geometry(
     path_centerline_lca="../data/centerline_lca.csv",
     bounding_sphere_radius_mm_rca=3.0,
     bounding_sphere_radius_mm_lca=3.0,
-    n_points_intramural=100,
-    anomalous_rca=True,
-    anomalous_lca=False,
+    n_points_takeoff_rca=100,
+    n_points_takeoff_lca=100,
+    acute_takeoff_rca=True,
+    acute_takeoff_lca=False,
     control_plot=True,
 )